separate hierCluster config code from the main hierCluster code
